Changed parent of tooltip to make it display.
Couldn't get it to display when I appended it to
the SVG chart so I've attached it to the body for
now so that I can see it in action.

If I keep the tooltip relative to each bar then
this might be fine but I may have to revert.