Changed variable name of `x1` in `timeline.html` to `t`.
This is to save confusion and having to remember that `x1`
refers to the non-fixed timeline in the chart. Used `t` for the
"t" in "content" (as in what is being focused on) as opposed to `x`
for the "x" in "context" which is fixed to show where in the
scale the content comes from.