Abstracted sample data more from timeline code.
Previously lanes were determined by number in the data and
the labelling was hardcoded into the timeline code. I've now
but the labels in the sample data where they belong (so one
can read and know the kingdoms each row refers to) and
sorted the ordering out in the code so that lane numbers
are calculated automatically.