Changed sample data and date-handling.
Now the years given in the sample date are full year format
with as many preceding zeros as required.

Date-handling is built around this idea.

The columns are also abstracted out in the commit so that one
could use different tsv files with different column names but
map one to the the standard chart dimensions in the code.